requirements.txt
openai==0.27.0
pinecone-client==2.2.0
langgraph==1.0.0
requests==2.28.1
PyYAML==6.0
serpapi==3.3.1
arxiv==0.2.0
